refactor: auth login and register
