+++ app/src/main/java/com/example/donutapptest/MainActivity.kt
@@ -8,7 +8,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
 import androidx.navigation.compose.rememberNavController
 import com.example.donutapptest.ui.navigation.NavigationComponent
 import com.example.donutapptest.ui.theme.DonutAppTestTheme
@@ -20,7 +19,10 @@ class MainActivity : ComponentActivity() {
         setContent {
             DonutAppTestTheme(dynamicColor = false) {
                 val navController = rememberNavController()
-                Scaffold(modifier = Modifier.fillMaxSize().padding(horizontal = 16.dp)) { innerPadding ->
+                Scaffold(
+                    modifier = Modifier
+                        .fillMaxSize()
+                ) { innerPadding ->
                     NavigationComponent(
                         navController = navController,
                         modifier = Modifier.padding(innerPadding)
