+++ app/src/main/java/com/example/donutapptest/ui/viewmodel/auth/RegisterViewModel.kt
@@ -1,0 +1,8 @@
+package com.example.donutapptest.ui.viewmodel.auth
+
+import androidx.lifecycle.ViewModel
+
+
+open class RegisterViewModel : ViewModel() {
+    
+}
